3634: Added role to MessageCardView
Currently, notifications are posted in a MessageCard element. Visually it is clear that the text (heading), the close button and the read more button are grouped because of the border.

People who cannot see the screen do not know that these elements are grouped. They only hear a heading, then ‘close, button’ (which gives no indication what exactly they are going to close) and then a ‘read more’ link. It is not clear they have encountered a message card that they can close.

Give the card surrounding the text and buttons an accessibilityLabel with a value of ‘notification' and add the word ‘notification' also to the close button: ‘close notification, button’ is expected.